# api/swagger.yaml
openapi: 3.0.1
info:
  title: MZT Health Fitness App
  description: A webapp to assist a coach in managing their client's workout programs.
  version: 0.0.1
servers:
- url: /
paths:
...
  /sessionTemplate/assignExercise:
    post:
      description: Assigns exercises to session template
      operationId: assignExercise
      requestBody:
        description: Assigns exercises to session template
        required: true
        content:
          application/json:
            schema:
               $ref: '#/components/parameters/exerciseIdParamAssign'
      responses:
        200:
          description: OK Session Template Info Obtained
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionTemplate'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: sessionTemplate.controller
...
  /sessionTemplate/getSessionTempById/{session_template_id}:
    get:
      description: Gets an session template by ID
      operationId: getSessionTempById
      parameters:
      - $ref: '#/components/parameters/sessionTempIdParamGet'
      responses:
        200:
          description: OK Session Info Obtained
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionTemplate'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: sessionTemplate.controller
  /sessionTemplate/addSessionTemp:
    post:
      description: Creates new session template
      operationId: addSessionTemp
      requestBody:
        description: Add a new session template
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionTemplate'
      responses:
        200:
          description: OK Session Info Obtained
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionTemplate'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: sessionTemplate.controller
  /sessionTemplate/updateSessionTemp:
    post:
      description: Updates an existing session template
      operationId: updateSessionTemp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionTemplate'
      responses:
        200:
          description: OK Session Template Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionTemplate'
                x-content-type: application/json
        400:
          description: 'Error: Bad Request'
        401:
          description: Unauthorized
        403:
          description: Forbidden
        404:
          description: 'Error: Not Found'
      security:
      - basicAuth: []
      x-swagger-router-controller: sessionTemplate.controller
...
components:
  parameters:
...
    exerciseIdParamAssign:
      name: assignExercise
      in: body
      description: the id of the exercise
      required: true
      style: simple
      explode: false
      schema:
        type: Array
        items: {}
...
    sessionTempIdParamGet:
        name: session_template_id
        in: path
        description: the id of the session template
        required: true
        style: simple
        explode: false
        schema:
          type: string
...
  schemas:
...
    SessionTemplate:
        type: object
        properties:
            session_template_id:
                type: object
            coach_id:
                type: object
            session_type:
                type: String
            session_coach_notes:
                type: String
            session_template_tag:
                type: array
                items: {}
            program_template_id:
                type: object